Moved landscape and chunk management to its own plugin
