+++ crates/world_server/examples/server.rs
@@ -1,7 +1,25 @@
-use projekto_world_server::{app, Landscape};
+use std::{thread, time::Duration};
+
+use projekto_world_server::{
+    app,
+    channel::{ServerCommand, ServerCommandSender},
+};
 
 fn main() {
     let mut server_app = app::new();
-    server_app.world.insert_resource(Landscape::default());
+
+    let sender = (*server_app.world.resource::<ServerCommandSender>()).clone();
+    thread::spawn(move || {
+        thread::sleep(Duration::from_millis(1000));
+        sender
+            .send(ServerCommand::LandscapeAdd(
+                Default::default(),
+                Default::default(),
+            ))
+            .unwrap();
+        thread::sleep(Duration::from_millis(1000));
+        sender.send(ServerCommand::LandscapeRemove).unwrap();
+    });
+
     server_app.run();
 }
